requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
dataclasses-json==0.5.13
filelock==3.9.0
frozenlist==1.4.0
fsspec==2023.6.0
greenlet==2.0.2
huggingface-hub==0.16.4
idna==3.4
Jinja2==3.1.2
langchain==0.0.242
langsmith==0.0.14
MarkupSafe==2.1.2
marshmallow==3.20.1
mpmath==1.2.1
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.0
numexpr==2.8.4
numpy==1.24.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
Pillow==9.3.0
PyAudio==0.2.13
pycparser==2.21
pydantic==1.10.12
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.6.3
requests==2.28.1
safetensors==0.3.1
scipy==1.11.1
SQLAlchemy==2.0.19
srt==3.5.3
sympy==1.11.1
tenacity==8.2.2
tokenizers==0.13.3
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
tqdm==4.65.0
transformers==4.31.0
typing-inspect==0.9.0
typing_extensions==4.4.0
urllib3==1.26.13
vosk==0.3.45
webrtcvad==2.0.10
websockets==11.0.3
yarl==1.9.2
